completed the False Positive post analysis and refactoring, completed the readme file and added notebook runs in the main.py
